[3307] Add Allocations update endpoint
PUT /api/v2/allocations/<id> to update an `Allocation`. Currently only
accepts a `number_of_places` parameter.
